# Add histogram validation