--- a/storage/interface.go
+++ b/storage/interface.go
@@ -27,14 +27,16 @@ import (
 
 // The errors exposed.
 var (
-	ErrNotFound                    = errors.New("not found")
-	ErrOutOfOrderSample            = errors.New("out of order sample")
-	ErrDuplicateSampleForTimestamp = errors.New("duplicate sample for timestamp")
-	ErrOutOfBounds                 = errors.New("out of bounds")
-	ErrOutOfOrderExemplar          = errors.New("out of order exemplar")
-	ErrDuplicateExemplar           = errors.New("duplicate exemplar")
-	ErrExemplarLabelLength         = fmt.Errorf("label length for exemplar exceeds maximum of %d UTF-8 characters", exemplar.ExemplarMaxLabelSetLength)
-	ErrExemplarsDisabled           = fmt.Errorf("exemplar storage is disabled or max exemplars is less than or equal to 0")
+	ErrNotFound                             = errors.New("not found")
+	ErrOutOfOrderSample                     = errors.New("out of order sample")
+	ErrDuplicateSampleForTimestamp          = errors.New("duplicate sample for timestamp")
+	ErrOutOfBounds                          = errors.New("out of bounds")
+	ErrOutOfOrderExemplar                   = errors.New("out of order exemplar")
+	ErrDuplicateExemplar                    = errors.New("duplicate exemplar")
+	ErrExemplarLabelLength                  = fmt.Errorf("label length for exemplar exceeds maximum of %d UTF-8 characters", exemplar.ExemplarMaxLabelSetLength)
+	ErrExemplarsDisabled                    = fmt.Errorf("exemplar storage is disabled or max exemplars is less than or equal to 0")
+	ErrHistogramSpanNegativeOffset          = errors.New("histogram has a span whose offset is negative")
+	ErrHistogramDifferentNumberSpansBuckets = errors.New("histogram spans specify different number of buckets than provided")
 )
 
 // SeriesRef is a generic series reference. In prometheus it is either a