--- a/tsdb/head_test.go
+++ b/tsdb/head_test.go
@@ -3805,7 +3805,7 @@ func TestReplayAfterMmapReplayError(t *testing.T) {
 	require.NoError(t, h.Close())
 }
 
-func TestDistributor_Push_HistogramValidation(t *testing.T) {
+func TestHistogramValidation(t *testing.T) {
 	tests := map[string]struct {
 		h      *histogram.Histogram
 		errMsg string
@@ -3818,48 +3818,48 @@ func TestDistributor_Push_HistogramValidation(t *testing.T) {
 				NegativeSpans:   []histogram.Span{{Offset: 0, Length: 1}},
 				NegativeBuckets: []int64{},
 			},
-			errMsg: `negative spans need 1 buckets, have 0 negative buckets`,
+			errMsg: `negative side: spans need 1 buckets, have 0 buckets`,
 		},
 		"rejects histogram who has too few positive buckets": {
 			h: &histogram.Histogram{
 				PositiveSpans:   []histogram.Span{{Offset: 0, Length: 1}},
 				PositiveBuckets: []int64{},
 			},
-			errMsg: `positive spans need 1 buckets, have 0 positive buckets`,
+			errMsg: `positive side: spans need 1 buckets, have 0 buckets`,
 		},
 		"rejects histogram who has too many negative buckets": {
 			h: &histogram.Histogram{
 				NegativeSpans:   []histogram.Span{{Offset: 0, Length: 1}},
 				NegativeBuckets: []int64{1, 2},
 			},
-			errMsg: `negative spans need 1 buckets, have 2 negative buckets`,
+			errMsg: `negative side: spans need 1 buckets, have 2 buckets`,
 		},
 		"rejects histogram who has too many positive buckets": {
 			h: &histogram.Histogram{
 				PositiveSpans:   []histogram.Span{{Offset: 0, Length: 1}},
 				PositiveBuckets: []int64{1, 2},
 			},
-			errMsg: `positive spans need 1 buckets, have 2 positive buckets`,
+			errMsg: `positive side: spans need 1 buckets, have 2 buckets`,
 		},
 		"rejects a histogram which has a negative span with a negative offset": {
 			h: &histogram.Histogram{
 				NegativeSpans:   []histogram.Span{{Offset: -1, Length: 1}, {Offset: -1, Length: 1}},
 				NegativeBuckets: []int64{1, 2},
 			},
-			errMsg: `negative span number 2 with offset -1`,
+			errMsg: `negative side: span number 2 with offset -1`,
 		},
 		"rejects a histogram which has a positive span with a negative offset": {
 			h: &histogram.Histogram{
 				PositiveSpans:   []histogram.Span{{Offset: -1, Length: 1}, {Offset: -1, Length: 1}},
 				PositiveBuckets: []int64{1, 2},
 			},
-			errMsg: `positive span number 2 with offset -1`,
+			errMsg: `positive side: span number 2 with offset -1`,
 		},
 	}
 
 	for testName, tc := range tests {
 		t.Run(testName, func(t *testing.T) {
-			err := validateHistogram(tc.h)
+			err := ValidateHistogram(tc.h)
 			if tc.errMsg != "" {
 				require.ErrorContains(t, err, tc.errMsg)
 			} else {