_This is to the sparsehistogram branch_

This PR implements the following validation checks:

1. Do any of the histogram's spans have a negative offset
2. Are there the appropriate number of negative and positive buckets based on their respective spans