@LeviHarrison 2022-07-28
Performance impact of the new checks looks pretty negligible I think (varies a little though between runs):

```
name                    old time/op    new time/op    delta
HistogramValidation-10    1.10µs ± 0%    1.54µs ± 0%   ~     (p=1.000 n=1+1)
```