@LeviHarrison 2022-07-28
Don't know why `TestSparseHistogram_Sum_AddOperator` is complaining about not getting the expected result now, the math checks out I think:

```diff
diff --git a/promql/engine_test.go b/promql/engine_test.go
index 472a43b16..1c6533527 100644
--- a/promql/engine_test.go
+++ b/promql/engine_test.go
@@ -3914,7 +3914,7 @@ func TestSparseHistogram_Sum_AddOperator(t *testing.T) {
                                },
                                {
                                        Schema:        0,
-                                       Count:         15,
+                                       Count:         36,
                                        Sum:           2345.6,
                                        ZeroThreshold: 0.001,
                                        ZeroCount:     5,
@@ -3933,7 +3933,7 @@ func TestSparseHistogram_Sum_AddOperator(t *testing.T) {
                                },
                                {
                                        Schema:        0,
-                                       Count:         15,
+                                       Count:         36,
                                        Sum:           1111.1,
                                        ZeroThreshold: 0.001,
                                        ZeroCount:     5,
@@ -3955,7 +3955,7 @@ func TestSparseHistogram_Sum_AddOperator(t *testing.T) {
                                Schema:        0,
                                ZeroThreshold: 0.001,
                                ZeroCount:     14,
-                               Count:         39,
+                               Count:         81,
                                Sum:           4691.2,
                                PositiveSpans: []histogram.Span{
                                        {Offset: 0, Length: 3},
```