@LeviHarrison 2022-07-29
> We are doing the error check wrongly here. If you check for err inside the loop, you may find out what is the problem.

Yup, another validation error. Fixed 👍🏻